@@ -36,30 +36,10 @@ func NewExecution(id, workflowID string) Execution {
 	return wrapExecution(domain.NewExecution(id, workflowID))
 }
 
-// NodeConfig holds the configuration for creating a new Node with UUID validation.
-type NodeConfig struct {
-	// ID is the node ID (will be validated as UUID)
-	ID string
-	// WorkflowID is the workflow ID this node belongs to (will be validated as UUID)
-	WorkflowID string
-	// Type is the node type (e.g., "http-request", "transform", "llm")
-	Type string
-	// Name is the display name for the node
-	Name string
-	// Config holds the node-specific configuration
-	Config map[string]any
-}
-
-// NewNode creates a new node.
-// Deprecated: Use NewNodeFromConfig for UUID validation.
-func NewNode(id, workflowID, nodeType, name string, config map[string]any) Node {
-	return domain.NewNode(id, workflowID, nodeType, name, config)
-}
-
 // NewNodeFromConfig creates a new node with UUID validation.
 // Returns an error if ID or WorkflowID are not valid UUIDs.
 func NewNodeFromConfig(cfg NodeConfig) (Node, error) {
-	return domain.NewNodeFromConfig(domain.NodeConfig{
+	return domain.NewNode(domain.NodeConfig{
 		ID:         cfg.ID,
 		WorkflowID: cfg.WorkflowID,
 		Type:       cfg.Type,
